requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
pymongo[srv]
fastapi
uvicorn
joblib
pydantic
python-dotenv
motor
langchain-openai